LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/get-animations-inactive-timeline-expected.txt


FAIL getAnimations includes inactive scroll-linked animations that have not been canceled assert_equals: Single running animation expected 1 but got 0
